+++ src/protocol1/msg.c
@@ -35,7 +35,8 @@ static int do_write(struct asfd *asfd,
 	{
 		if((ret=bfd->write(bfd, out, outlen))<=0)
 		{
-			logp("error when appending %zu: %d\n", outlen, ret);
+			logp("error when appending %lu: %d\n",
+				(unsigned long)outlen, ret);
 			asfd->write_str(asfd, CMD_ERROR, "write failed");
 			return -1;
 		}
