+++ src/protocol1/rs_buf.c
@@ -107,8 +107,9 @@ rs_result rs_infilebuf_fill(rs_job_t *job, rs_buffers_t *buf, void *opaque)
 		//	buf->avail_in, fb->buf_len);
 		if(buf->avail_in > fb->buf_len)
 		{
-			logp("buf->avail_in > fb->buf_len (%zu > %zu) in %s\n",
-				buf->avail_in, fb->buf_len, __func__);
+			logp("buf->avail_in > fb->buf_len (%lu > %lu) in %s\n",
+				(unsigned long)buf->avail_in,
+				(unsigned long)fb->buf_len, __func__);
 			return RS_IO_ERROR;
 		}
 		if(buf->next_in < fb->buf)
@@ -127,8 +128,8 @@ rs_result rs_infilebuf_fill(rs_job_t *job, rs_buffers_t *buf, void *opaque)
 	{
 		if(buf->avail_in)
 		{
-			logp("buf->avail_in is %zu in %s\n",
-				buf->avail_in, __func__);
+			logp("buf->avail_in is %lu in %s\n",
+				(unsigned long)buf->avail_in, __func__);
 			return RS_IO_ERROR;
 		}
 	}
@@ -248,8 +249,8 @@ rs_result rs_outfilebuf_drain(rs_job_t *job, rs_buffers_t *buf, void *opaque)
 	{
 		if(buf->avail_out)
 		{
-			logp("buf->avail_out is %zu in %s\n",
-				buf->avail_out, __func__);
+			logp("buf->avail_out is %lu in %s\n",
+				(unsigned long)buf->avail_out, __func__);
 			return RS_IO_ERROR;
 		}
 		buf->next_out = fb->buf;
@@ -259,8 +260,9 @@ rs_result rs_outfilebuf_drain(rs_job_t *job, rs_buffers_t *buf, void *opaque)
 
 	if(buf->avail_out > fb->buf_len)
 	{
-		logp("buf->avail_out > fb->buf_len (%zu > %zu) in %s\n",
-			buf->avail_out, fb->buf_len, __func__);
+		logp("buf->avail_out > fb->buf_len (%lu > %lu) in %s\n",
+			(unsigned long)buf->avail_out,
+			(unsigned long)fb->buf_len, __func__);
 		return RS_IO_ERROR;
 	}
 	if(buf->next_out < fb->buf)
