+++ src/protocol2/blk.c
@@ -112,7 +112,8 @@ int blk_set_from_iobuf_sig(struct blk *blk, struct iobuf *iobuf)
 {
 	if(iobuf->len!=24)
 	{
-		logp("Signature wrong length: %zu!=24\n", iobuf->len);
+		logp("Signature wrong length: %lu!=24\n",
+			(unsigned long)iobuf->len);
 		return -1;
 	}
 	set_sig(blk, iobuf);
@@ -123,8 +124,8 @@ int blk_set_from_iobuf_sig_and_savepath(struct blk *blk, struct iobuf *iobuf)
 {
 	if(iobuf->len!=32)
 	{
-		logp("Signature with save_path wrong length: %zu!=32\n",
-			iobuf->len);
+		logp("Signature with save_path wrong length: %lu!=32\n",
+			(unsigned long)iobuf->len);
 		return -1;
 	}
 	set_sig(blk, iobuf);
@@ -136,8 +137,9 @@ int blk_set_from_iobuf_fingerprint(struct blk *blk, struct iobuf *iobuf)
 {
 	if(iobuf->len!=sizeof(blk->fingerprint))
 	{
-		logp("Fingerprint wrong length: %zu!=%zu\n",
-			iobuf->len, sizeof(blk->fingerprint));
+		logp("Fingerprint wrong length: %lu!=%lu\n",
+			(unsigned long)iobuf->len,
+			(unsigned long)sizeof(blk->fingerprint));
 		return -1;
 	}
 	set_fingerprint(blk, iobuf);
@@ -148,8 +150,9 @@ int blk_set_from_iobuf_savepath(struct blk *blk, struct iobuf *iobuf)
 {
 	if(iobuf->len!=sizeof(blk->savepath))
 	{
-		logp("Save path wrong length: %zu!=%zu\n",
-			iobuf->len, sizeof(blk->savepath));
+		logp("Save path wrong length: %lu!=%lu\n",
+			(unsigned long)iobuf->len,
+			(unsigned long)sizeof(blk->savepath));
 		return -1;
 	}
 	set_savepath(blk, iobuf, 0 /* offset */);
@@ -160,8 +163,8 @@ int blk_set_from_iobuf_index_and_savepath(struct blk *blk, struct iobuf *iobuf)
 {
 	if(iobuf->len!=16)
 	{
-		logp("File number and savepath with wrong length: %zu!=16\n",
-			iobuf->len);
+		logp("File number and savepath with wrong length: %lu!=16\n",
+			(unsigned long)iobuf->len);
 		return -1;
 	}
 	blk->index=ETOH(*(uint64_t *)iobuf->buf);
@@ -173,8 +176,9 @@ int blk_set_from_iobuf_wrap_up(struct blk *blk, struct iobuf *iobuf)
 {
 	if(iobuf->len!=sizeof(blk->index))
 	{
-		logp("Wrap up with wrong length: %zu!=%zu\n",
-			iobuf->len, sizeof(blk->index));
+		logp("Wrap up with wrong length: %lu!=%lu\n",
+			(unsigned long)iobuf->len,
+			(unsigned long)sizeof(blk->index));
 		return -1;
 	}
 	blk->index=ETOH(*(uint64_t *)iobuf->buf);
